Add SignupFormHeader component and mixins for common font sizes
