+++ client/landing/gutenboarding/mixins.scss
@@ -3,3 +3,35 @@
 	font-weight: 400;
 	letter-spacing: -0.4px;
 }
+
+@mixin onboarding-heading-text {
+	@include onboarding-font-recoleta;
+	font-size: 42px;
+	line-height: 57px;
+}
+
+@mixin onboarding-heading-text-mobile {
+	@include onboarding-font-recoleta;
+	font-size: 32px;
+	line-height: 40px;
+}
+
+@mixin onboarding-large-text {
+	font-size: 16px;
+	line-height: 19px;
+}
+
+@mixin onboarding-medium-text {
+	font-size: 14px;
+	line-height: 17px;
+}
+
+@mixin onboarding-small-text {
+	font-size: 12px;
+	line-height: 14px;
+}
+
+@mixin onboarding-x-small-text {
+	font-size: 10px;
+	line-height: 12px;
+}
